processor/requirements.txt
python-dotenv
pymongo
psycopg2-binary
boto3
